Correct Helm deployment selector.matchLabels
If selector.matchLabels is not set, Kubernetes automatically copies
deployment template labels to selector.matchLabels. In this case,
including the version in the helm.sh/chart label, as recommended in
Helm best practices, causes upgrades to fail because
selector.matchLabels is immutable. To resolve this, specify matchLabels
by hand, excluding the chart label.